% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asociaciones_Gajardo.R
\docType{data}
\name{Asociaciones_Gajardo}
\alias{Asociaciones_Gajardo}
\title{Asociaciones Gajardo}
\format{
## `Asociaciones_Gajardo`
A tibble with 232 rows and 4 columns:
\describe{
  \item{REGIÓN, SUB-REGIÓN, FORMACIÓN}{Detalle de las Formaciones a diferentes escalas}
  \item{ASOCIACIÓN}{Asociaciones vegetales presentes en cada formación}
}
}
\source{
Elaboración propia a partir de Gajardo (1994)
}
\usage{
Asociaciones_Gajardo
}
\description{
Asociaciones por cada formación vegetacional descrita por Gajardo (1994)
}
\keyword{datasets}
