% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Especies_Gajardo.R
\docType{data}
\name{Especies_Gajardo}
\alias{Especies_Gajardo}
\title{Especies Gajardo}
\format{
## `Especies_Gajardo`
A tibble with 605 rows and 4 columns:
\describe{
  \item{ASOCIACIÓN}{Asociaciones vegetales}
  \item{TIPO_ESPECIE}{Representativa, Acompañante, Común u Ocacional}
  \item{ESPECIE, NOMBRE.COMÚN}{Nombre científico y común de la espcies}
}
}
\source{
Elaboración propia a partir de Gajardo (1994)
}
\usage{
Especies_Gajardo
}
\description{
Especies que componen las asociaciones vegetales descrita por Gajardo (1994)
}
\keyword{datasets}
