% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_RedVial.R
\name{Get_RedVial}
\alias{Get_RedVial}
\title{Capas vectoriales Red Vial}
\usage{
Get_RedVial(shp = NULL, var = NULL, val = NULL)
}
\arguments{
\item{shp}{Objeto sf.}

\item{var}{Campo que desea filtrar. Los campos disponibles son "CUT_REG", "CUT_PROV", "CUT_COM", "REGION", "PROVINCIA" y "COMUNA".}

\item{val}{Valores del campo seleccionado por los que se desea filtrar.}
}
\value{
Capa vectorial en formato sf. CRS: EPSG: 4326. Unit: degree
}
\description{
Capas vectoriales Red Vial
}
\examples{
Get_RedVial(var = "ROL", val = "IPA 15")
Get_RedVial(var = "CARPETA", val = c("Pavimento Doble Calzada","Grava Tratada"))

}
