% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LyP.R
\docType{data}
\name{LyP}
\alias{LyP}
\title{Pisos vegetacionales}
\format{
## `LyP`
Simple feature collection with 126 features and 9 fields:
\describe{
  \item{codigo}{codigo del piso vegetacional}
  \item{piso}{nombre del piso vegetacional}
  \item{formacion}{formacion a la que corresponde el piso vegetacional}
  \item{Area_km2, Length_km}{Área y longitud del piso vegetacional}
}
}
\source{
<http://www.ide.cl/descargas/capas/MMA/PisosVegetacionalesPliscoff2017.zip>
}
\usage{
LyP
}
\description{
Pisos vegetacionales descritos por Luebert y Pliscoff (2017)
}
\keyword{datasets}
