% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formaciones_Gajardo.R
\docType{data}
\name{Formaciones_Gajardo}
\alias{Formaciones_Gajardo}
\title{Formaciones Gajardo}
\format{
## `Formaciones_Gajardo`
Simple feature collection with 939 features and 8 fields:
\describe{
  \item{Formacion, Subregion, Region, Distrito, Provincia}{Detalle de las Formaciones a diferentes escalas}
  \item{Sup_km2, Leng_km}{Área y longitud del polígono de la formación}
}
}
\source{
<https://www.ide.cl/index.php/flora-y-fauna/item/1515-formaciones-vegetacionales-gajardo>
}
\usage{
Formaciones_Gajardo
}
\description{
Formaciones vegetales descritas por Gajardo (1994)
}
\keyword{datasets}
